Moved DBVector to into separate module...
